* 'master' of https://github.com/IsabelChong/tp: (46 commits)
  fix checkstyle
  Adds getter for ObservableList
  Implemented viewing attendance based on student - `attendance` under Student Page will show number of sessions attended
  Remove outdated fields in `Student.java` - Tags and courses
  Allow UploadCommand only in students page
  Fix checkstyle errors
  Fixed GUI bug for student view pane
  Rebase cleanup
  Optimise clippy
  Fix checkstyle
  Change find command to display full list if there are no results
  Fix remark command
  Clear filters when navigating between pages
  Optimisation for UI navigation handling
  Change commands to perform on FilteredList instead of internalList
  Change UI to show FilteredList instead of internalList
  Fix wrong page restriction for find student
  Add page restrictions for find command types
  Search keywords using .contains() instead of by word
  Implement find command
  ...

+++ src/main/java/tfifteenfour/clipboard/logic/commands/UploadCommand.java
@@ -3,12 +3,14 @@ package tfifteenfour.clipboard.logic.commands;
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 import static java.util.Objects.requireNonNull;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import tfifteenfour.clipboard.logic.CurrentSelection;
+import tfifteenfour.clipboard.logic.PageType;
 import tfifteenfour.clipboard.logic.commands.exceptions.CommandException;
 import tfifteenfour.clipboard.model.Model;
 
@@ -58,9 +60,18 @@ public class UploadCommand extends Command {
 
     @Override
     public CommandResult execute(Model model, CurrentSelection currentSelection) throws CommandException {
+
+        if (currentSelection.getCurrentPage() != PageType.STUDENT_PAGE) {
+            throw new CommandException("Wrong page. Navigate to student page to upload a file.");
+        }
+
         try {
             Path sourcePath = this.sourcePath;
             Path destPath = this.destPath;
+            File toCopy = new File(String.valueOf(sourcePath));
+            if (!toCopy.isFile()) {
+                throw new CommandException("Please upload a valid file.");
+            }
             Files.copy(sourcePath, destPath.resolve(sourcePath.getFileName()), REPLACE_EXISTING);
             return new CommandResult(this, generateSuccessMessage(sourcePath), willModifyState);
         } catch (IOException e) {
